backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Text processing and chunking
langchain==0.0.335
tiktoken==0.5.2

# PostgreSQL with vector support
psycopg[binary]==3.1.18
pgvector==0.2.4
sqlalchemy==2.0.41
alembic==1.13.1

# OpenAI API
openai==1.3.6

# Additional utilities
python-dotenv==1.0.0
# Updated for Python 3.13 compatibility
pandas>=2.2.0
numpy>=1.26.0
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: for better performance
python-json-logger==2.0.7 